Splits apart the line of data in the file
Updated it so that the line of data in the file will be split into two returnable values: the ID and the time stamp. Then is a nascent validation check it determines whether or not the IDs match. If they do match, it returns both the ID and the time stamp.